requirements.txt
# Core ML and Data Science Libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.10.0

# Advanced ML Libraries
xgboost>=1.7.0
lightgbm>=4.0.0
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
evaluate>=0.4.0

# Visualization Libraries
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
bokeh>=3.2.0

# Web Interface
streamlit>=1.25.0
gradio>=3.40.0

# Data Processing
openpyxl>=3.1.0
xlrd>=2.0.0
pyarrow>=12.0.0

# Configuration and Logging
pyyaml>=6.0
loguru>=0.7.0
python-dotenv>=1.0.0

# Performance and Optimization
numba>=0.57.0
joblib>=1.3.0
dask>=2023.6.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0
pre-commit>=3.3.0

# Jupyter and Interactive Development
jupyter>=1.0.0
ipython>=8.14.0
notebook>=6.5.0

# Optional: Deep Learning Extras
# tensorflow>=2.13.0
# keras>=2.13.0

# Optional: Time Series
# statsmodels>=0.14.0
# prophet>=1.1.0

# Optional: Geospatial Analysis
# geopandas>=0.13.0
# folium>=0.14.0